refactor: separate handleKeyUp logic from useTextarea
